requirements.txt
streamlit==1.37.1
beautifulsoup4 == 4.12.3
pillow==10.4.0
pandas==2.2.2
requests==2.32.3
openpyxl==3.1.5
lxml==5.3.0